backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
attrs==25.4.0
azure-cognitiveservices-speech==1.36.0
azure-common==1.1.28
azure-core==1.38.0
azure-identity==1.15.0
azure-search-documents==11.4.0
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.109.0
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
ics==0.7.2
idna==3.11
isodate==0.7.2
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-openai==1.1.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
Mako==1.3.10
MarkupSafe==3.0.3
msal==1.34.0
msal-extensions==1.3.1
oauthlib==3.3.1
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
passlib==1.7.4
proto-plus==1.27.0
protobuf==6.33.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
starlette==0.35.1
TatSu==5.16.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
